refactor(ui/web/sponsorship): rename files
